+++ test/clj_kondo/impl/types_test.clj
@@ -46,6 +46,10 @@
   (testing "fix for #1023"
     (is (= #{} (types-utils/union-type)))))
 
+(deftest lenient-match-test
+  (is (types/match? ::unknown-type :string))
+  (is (types/match? :string :nilable/any)))
+
 ;;;; Scratch
 
 (comment